+++ Changelog.md
@@ -6,6 +6,7 @@
 - Fix issue with `coq.env.add-abbreviation` when given a term with binders
   having overlapping `name`s.
 - New `copy-indt-decl` 
+- New `coq.coercion.declare` is able to infer the endpoints if omitted
 
 ## [1.9.3] - 18-02-2021